file-operations: Batch process conflicts
Currently, nautilus deals with file and folder conflicts individually, prompting
the user to make a decision on each conflict operation in sequence. The proposed
file conflict dialog redesign envisions multiple conflicts being collated into a
single dialog for the user to act on.

To make this possible, teach file-operations to store each conflict it encounters
for processing in a batch that can be passed to operations-ui-manager to populate
the multi-file conflict dialog.