file-operations: use AdwMessageDialog
Add a AdwMessageDialog-based helper function to present a dialog with
responses. It supports showing an additional text, along the heading and
body texts, so it can be used as a replacement for file operations
dialogs.

A response enum is used to indicate which responses to present and which
response was chosen. The enum values can be combined as bit flags to
represent multiple responses. The order of responses is thus fixed.
Responses that represent an "All" action have to be passed explicitly,
making it clear when such responses will be presented.

`run_dialog()` also returns a response enum value, allowing for more
verbose response comparisons.

Fixes #2900