properties-window: Tweak the "Others" section design
The insensitive "Others" is only present to give the a title
to the "Access" row. However, this is a misleading design,
as users might expect that the "Others" row would be
triggerable in some condition.

Also, rename "Others" to "Other Users", as per recommendation
from https://gitlab.gnome.org/GNOME/nautilus/-/merge_requests/1499#note_2099507