window-slot: Do not force reload on same location when error is present
When the user enters a privileged directory and reloads after the polkit authentication has expired, they are unable to cancel the dialog as it continously pops up again.

This is because we are forcing a slot reload on error, but since we are going to the same location for a reload the user will be reprompted with the authentication dialog.

Check if the pending location is the same as the current slot location and only force a reload if they are different.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3481