search-popover: Add search options functionality
This commit introduces code to make the "Only Filenames" and "Only
Contents" buttons functional. Two variables are introduced: fts_enabled
and fts_sensitive. The fts_sensitive variable sets the search options'
sensitivity (i.e., whether they can be selected by the user), such as
when viewing a remote location. The fts_enabled variable checks which
button is clicked: when set to 1, "Only Filenames" is disabled; when set
to 0, "Only Contents" is disabled.

Additionally, a property named 'fts-enabled' is installed in the
popover, which is notified when either of the buttons is clicked. When
notified, a callback in the nautilus-query-editor is called, updating
the query-editor to add a tag in the search field.