search-popover: Add functionality for file type buttons
This commit introduces code to make the file type buttons functional.
First, a key-value pair is added to each file type button. The key is
"mimetype-group," and the value is an integer representing the position
of that mimetype in the mimetype_groups array in
nautilus-mime-actions.c. When one of these buttons is clicked, a signal
is emitted, and a callback is invoked in the query editor. This callback
takes the button name and the mimetype group and creates a tag in the
query editor.

In addition to the visible buttons, there is an "other file types"
dialog box that appears when the "other file types" button is clicked.
The code for this dialog box is directly copied from the previous
implementation.