search-popover: Add functionality for date buttons
This commit introduces the code to make the date buttons functional.
Similar to the file type buttons, a key-value pair is added to the
visible date buttons where the key is "date" and the value is a datetime
with the number of days subtracted as indicated on the button. When the
date buttons are clicked, a date range pointer array is created where
the first pointer is the older date and the second pointer is today's
date. This array is passed to the callback function, where a new tag is
created for the search field.