preferences: Clear weak pointer
We are using a weak pointer to make sure that there's never more
than 1 preference window showing at a time.  Unfortunately, it
appears that the preference window is being finalized at the
wrong time, and clearing the weak pointer after it's been set
to a new window.  Instead make sure that the weak pointer is
cleared when it should be.

Fixes: https://gitlab.gnome.org/GNOME/nautilus/-/issues/3576