style: Fix AdwTabOverview tab thumbnails having glitchy border corners
Since the .view style class is set directly in
NautilusWindow, the .view .thumbnail style class also
affects AdwTabOverview, which is a child of NautilusWindow.
This causes the AdwTabOverview tab thumbnails to have
glitchy border corners because the checkerboard pattern is
also being applied to them.

Use the .nautilus-grid/list-view style class instead.

Co-authored-by: Alice Mikhaylenko <alicem@gnome.org>